Added casts back to Record init with Builder only for primitive (nonreference) types, added thorough record PDSC

+++ restlicsharpdata/restlidataintegration/com/linkedin/restli/datagenerator/integration/RecordInUnionTest.cs
@@ -32,7 +32,7 @@ namespace com.linkedin.restli.datagenerator.integration
       // Retrieve data for a
       if (builder.a != null)
       {
-        a = builder.a;
+        a = (int)builder.a;
 
       }
       else
